requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart